+++ packages/backend-split/src/experimental/features.http
@@ -1,6 +1,7 @@
-// Not working yet, the metadata plugin HTTP APIs need to be moved to rootHttpRouter.
-# GET http://localhost:7007/.backstage/systemInfo/features/installed
+// Test requests for the split backend to test the system metadata service :)
+// The 2 below requests should return the same data.
+GET http://localhost:7007/.backstage/systemMetadata/v1/features/installed
 
 ###
 
-GET http://localhost:7008/.backstage/systemInfo/features/installed
+GET http://localhost:7008/.backstage/systemMetadata/v1/features/installed
